All LLM Tests pass

+++ src/relax/frontend/nn/llm/kv_cache_attention_decode_gpu_helpers.cc
@@ -54,7 +54,8 @@ DecodeGpuConfig ComputeDecodeGpuConfig(int64_t H_kv, int64_t H_qo, int64_t D,
     TILE_SIZE_PER_BDX = 1;
   }
 
-  int64_t VEC_SIZE = std::min(std::max(8LL / qkv_dtype_bytes, D / 32), 4LL);
+  int64_t VEC_SIZE = std::min(std::max(static_cast<int64_t>(8LL / qkv_dtype_bytes), D / 32),
+                              static_cast<int64_t>(4LL));
 
   // Query max threads per block from target attrs
   int64_t max_threads = 1024;
@@ -82,9 +83,9 @@ DecodeGpuConfig ComputeDecodeGpuConfig(int64_t H_kv, int64_t H_qo, int64_t D,
 // ============================================================================
 
 PrimExpr BuildDecodeRopeExpr(tir::Buffer buf, ffi::Array<PrimExpr> base_indices, PrimExpr d_expr,
-                              int64_t D, PrimExpr pos_expr, tir::Var rope_scale,
-                              tir::Var rope_theta, tir::Var rotary_mode, const std::string& dtype,
-                              const ffi::Map<ffi::String, ffi::Any>& rope_scaling) {
+                             int64_t D, PrimExpr pos_expr, tir::Var rope_scale, tir::Var rope_theta,
+                             tir::Var rotary_mode, const std::string& dtype,
+                             const ffi::Map<ffi::String, ffi::Any>& rope_scaling) {
   bool is_f16 = (dtype == "float16");
   DataType dt = DataType(runtime::StringToDLDataType(dtype));
   int64_t half_d = D / 2;
@@ -100,9 +101,9 @@ PrimExpr BuildDecodeRopeExpr(tir::Buffer buf, ffi::Array<PrimExpr> base_indices,
   ffi::Array<PrimExpr> idx_minus = base_indices;
   idx_minus.push_back(d_expr - I32(half_d));
   PrimExpr neg_one = tir::make_const(dt, -1.0);
-  PrimExpr partner = tvm::if_then_else(d_expr < I32(half_d),
-                                        tir::BufferLoad(buf, idx_plus) * neg_one,
-                                        tir::BufferLoad(buf, idx_minus));
+  PrimExpr partner =
+      tvm::if_then_else(d_expr < I32(half_d), tir::BufferLoad(buf, idx_plus) * neg_one,
+                        tir::BufferLoad(buf, idx_minus));
 
   // Rope frequency: use d_expr directly (no extra d_var binding, matches Python TVMScript)
   RopeFreqFunc rope_freq_func = SwitchRopeFreqFunc(rope_scaling);
@@ -114,9 +115,9 @@ PrimExpr BuildDecodeRopeExpr(tir::Buffer buf, ffi::Array<PrimExpr> base_indices,
   // For f16: cos_freq/sin_freq are already float32 (from tvm::cos/sin), so no extra cast needed
   PrimExpr rope_val;
   if (is_f16) {
-    rope_val = CastTo(freq.cos_freq * CastTo(elem, "float32") +
-                      freq.sin_freq * CastTo(partner, "float32"),
-                      dtype);
+    rope_val =
+        CastTo(freq.cos_freq * CastTo(elem, "float32") + freq.sin_freq * CastTo(partner, "float32"),
+               dtype);
   } else {
     rope_val = freq.cos_freq * elem + freq.sin_freq * partner;
   }
@@ -160,11 +161,10 @@ Stmt BuildDecodeKVLoadBody(const DecodeGpuConfig& cfg, tir::Buffer pages_buf,
   tir::SizeVar seq_offset("seq_offset", DataType::Int(32));
   PrimExpr seq_offset_expr;
   if (sliding_window) {
-    seq_offset_expr = tvm::if_then_else(
-        row_g < tir::BufferLoad(length_info_buf, {I32(2), batch_idx}),
-        row_g,
-        row_g - tir::BufferLoad(length_info_buf, {I32(2), batch_idx}) +
-            tir::BufferLoad(length_info_buf, {I32(1), batch_idx}));
+    seq_offset_expr =
+        tvm::if_then_else(row_g < tir::BufferLoad(length_info_buf, {I32(2), batch_idx}), row_g,
+                          row_g - tir::BufferLoad(length_info_buf, {I32(2), batch_idx}) +
+                              tir::BufferLoad(length_info_buf, {I32(1), batch_idx}));
   } else {
     seq_offset_expr = row_g;
   }
@@ -178,10 +178,10 @@ Stmt BuildDecodeKVLoadBody(const DecodeGpuConfig& cfg, tir::Buffer pages_buf,
   PrimExpr k_pos = tir::BufferLoad(k_rope_pos_offset_buf, {batch_idx}) + row_g;
   DataType pages_dt = pages_buf->dtype;
   std::string pages_dtype_str = ffi::DLDataTypeToString(pages_dt);
-  PrimExpr k_rope = BuildDecodeRopeExpr(
-      pages_buf, ffi::Array<PrimExpr>{page_no, I32(0), by_var, page_offset_var},
-      d_kv, pages_buf->shape[4].as<IntImmNode>()->value,
-      k_pos, rope_scale, rope_theta, rotary_mode, pages_dtype_str, rope_scaling);
+  PrimExpr k_rope =
+      BuildDecodeRopeExpr(pages_buf, ffi::Array<PrimExpr>{page_no, I32(0), by_var, page_offset_var},
+                          d_kv, pages_buf->shape[4].as<IntImmNode>()->value, k_pos, rope_scale,
+                          rope_theta, rotary_mode, pages_dtype_str, rope_scaling);
   PrimExpr v_elem = tir::BufferLoad(pages_buf, {page_no, I32(1), by_var, page_offset_var, d_kv});
 
   PrimExpr zero_kv = tir::make_const(K_smem->dtype, 0.0);
@@ -191,27 +191,24 @@ Stmt BuildDecodeKVLoadBody(const DecodeGpuConfig& cfg, tir::Buffer pages_buf,
       seq_offset, seq_offset_expr,
       tir::LetStmt(
           page_no,
-          tir::BufferLoad(page_values_buf,
-                          {cur_begin + floordiv(seq_offset, I32(page_size))}),
+          tir::BufferLoad(page_values_buf, {cur_begin + floordiv(seq_offset, I32(page_size))}),
           tir::LetStmt(
               page_offset_var, floormod(seq_offset, I32(page_size)),
               tir::For(vec, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
-                       tir::SeqStmt({
-                           tir::BufferStore(K_smem, k_rope, {tile_start_s + j_var, d_kv}),
-                           tir::BufferStore(V_smem, v_elem, {tile_start_s + j_var, d_kv})})))));
+                       tir::SeqStmt(
+                           {tir::BufferStore(K_smem, k_rope, {tile_start_s + j_var, d_kv}),
+                            tir::BufferStore(V_smem, v_elem, {tile_start_s + j_var, d_kv})})))));
 
   // Build the "row_g >= kv_len" zero-fill branch: single vec loop for both K and V
   tir::Var vec0("vec", DataType::Int(32));
   PrimExpr d_zero = tx * I32(cfg.VEC_SIZE) + vec0;
-  Stmt zero_branch = tir::For(
-      vec0, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
-      tir::SeqStmt({
-          tir::BufferStore(K_smem, zero_kv, {tile_start_s + j_var, d_zero}),
-          tir::BufferStore(V_smem, zero_kv, {tile_start_s + j_var, d_zero})}));
+  Stmt zero_branch =
+      tir::For(vec0, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
+               tir::SeqStmt({tir::BufferStore(K_smem, zero_kv, {tile_start_s + j_var, d_zero}),
+                             tir::BufferStore(V_smem, zero_kv, {tile_start_s + j_var, d_zero})}));
 
   // Wrap in row_g LetStmt
-  return tir::LetStmt(row_g, row_g_expr,
-                      tir::IfThenElse(row_g < kv_len, load_branch, zero_branch));
+  return tir::LetStmt(row_g, row_g_expr, tir::IfThenElse(row_g < kv_len, load_branch, zero_branch));
 }
 
 // ============================================================================
@@ -232,50 +229,46 @@ Stmt BuildDecodeAllreduceBlock(const DecodeGpuConfig& cfg, tir::Buffer O_allredu
   // Store O_local → O_allreduce[tz, ty, tx*VEC_SIZE+vec]
   tir::Var vec_store("vec", DataType::Int(32));
   Stmt store_O = tir::For(vec_store, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
-                           tir::BufferStore(O_allreduce, ldi(O_local, vec_store),
-                                            {tz, ty, tx * I32(cfg.VEC_SIZE) + vec_store}));
+                          tir::BufferStore(O_allreduce, ldi(O_local, vec_store),
+                                           {tz, ty, tx * I32(cfg.VEC_SIZE) + vec_store}));
 
   // Store st_m, st_d → md_allreduce[tz, ty, 0/1]
-  Stmt store_md = tir::SeqStmt({
-      tir::BufferStore(md_allreduce, ld0(st_m), {tz, ty, I32(0)}),
-      tir::BufferStore(md_allreduce, ld0(st_d), {tz, ty, I32(1)})});
+  Stmt store_md = tir::SeqStmt({tir::BufferStore(md_allreduce, ld0(st_m), {tz, ty, I32(0)}),
+                                tir::BufferStore(md_allreduce, ld0(st_d), {tz, ty, I32(1)})});
 
   Stmt sync = tir::Evaluate(tir::Call(DataType::Int(32), tir::builtin::tvm_storage_sync(),
-                                       ffi::Array<PrimExpr>{tir::StringImm("shared")}));
+                                      ffi::Array<PrimExpr>{tir::StringImm("shared")}));
 
   // Reset st_m, st_d, O_local
   tir::Var vec_reset("vec", DataType::Int(32));
-  Stmt reset = tir::SeqStmt({
-      st0(st_m, F32(-50000.0)),
-      st0(st_d, F32(1.0)),
-      tir::For(vec_reset, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
-               sti(O_local, vec_reset, F32(0.0)))});
+  Stmt reset =
+      tir::SeqStmt({st0(st_m, F32(-50000.0)), st0(st_d, F32(1.0)),
+                    tir::For(vec_reset, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
+                             sti(O_local, vec_reset, F32(0.0)))});
 
   // Merge loop: for j in range(bdz)
   tir::Var j_merge("j", DataType::Int(32));
   tir::Var vec_oo("vec", DataType::Int(32));
   tir::Var vec_om("vec", DataType::Int(32));
 
-  Stmt load_other_o = tir::For(
-      vec_oo, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
-      sti(other_o, vec_oo,
-          tir::BufferLoad(O_allreduce, {j_merge, ty, tx * I32(cfg.VEC_SIZE) + vec_oo})));
+  Stmt load_other_o =
+      tir::For(vec_oo, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
+               sti(other_o, vec_oo,
+                   tir::BufferLoad(O_allreduce, {j_merge, ty, tx * I32(cfg.VEC_SIZE) + vec_oo})));
 
-  Stmt merge_body = tir::SeqStmt({
-      st0(m_prev, ld0(st_m)),
-      st0(d_prev, ld0(st_d)),
-      st0(other_m, tir::BufferLoad(md_allreduce, {j_merge, ty, I32(0)})),
-      st0(other_d, tir::BufferLoad(md_allreduce, {j_merge, ty, I32(1)})),
-      load_other_o,
-      st0(st_m, tvm::max(ld0(st_m), ld0(other_m))),
-      st0(st_d, ld0(d_prev) * tvm::exp2(ld0(m_prev) - ld0(st_m)) +
-                ld0(other_d) * tvm::exp2(ld0(other_m) - ld0(st_m))),
-      st0(exp_mprev, tvm::exp2(ld0(m_prev) - ld0(st_m))),
-      st0(exp_otherm, tvm::exp2(ld0(other_m) - ld0(st_m))),
-      tir::For(vec_om, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
-               sti(O_local, vec_om,
-                   ldi(O_local, vec_om) * ld0(exp_mprev) +
-                   ldi(other_o, vec_om) * ld0(exp_otherm)))});
+  Stmt merge_body =
+      tir::SeqStmt({st0(m_prev, ld0(st_m)), st0(d_prev, ld0(st_d)),
+                    st0(other_m, tir::BufferLoad(md_allreduce, {j_merge, ty, I32(0)})),
+                    st0(other_d, tir::BufferLoad(md_allreduce, {j_merge, ty, I32(1)})),
+                    load_other_o, st0(st_m, tvm::max(ld0(st_m), ld0(other_m))),
+                    st0(st_d, ld0(d_prev) * tvm::exp2(ld0(m_prev) - ld0(st_m)) +
+                                  ld0(other_d) * tvm::exp2(ld0(other_m) - ld0(st_m))),
+                    st0(exp_mprev, tvm::exp2(ld0(m_prev) - ld0(st_m))),
+                    st0(exp_otherm, tvm::exp2(ld0(other_m) - ld0(st_m))),
+                    tir::For(vec_om, I32(0), I32(cfg.VEC_SIZE), tir::ForKind::kVectorized,
+                             sti(O_local, vec_om,
+                                 ldi(O_local, vec_om) * ld0(exp_mprev) +
+                                     ldi(other_o, vec_om) * ld0(exp_otherm)))});
 
   Stmt merge_loop = tir::For(j_merge, I32(0), I32(cfg.bdz), tir::ForKind::kSerial, merge_body);
